Fix some small bug in FontEnd

+++ KidProjectServer/KidProjectServer/Controllers/MenuController.cs
@@ -88,10 +88,13 @@ namespace KidProjectServer.Controllers
                     await formData.Image.CopyToAsync(stream);
                 }
                 // Delete old image if it exists
-                var oldImagePath = Path.Combine(_configuration["ImagePath"], oldMenu.Image);
-                if (System.IO.File.Exists(oldImagePath))
+                if (oldMenu.Image != null)
                 {
-                    System.IO.File.Delete(oldImagePath);
+                    var oldImagePath = Path.Combine(_configuration["ImagePath"], oldMenu.Image);
+                    if (System.IO.File.Exists(oldImagePath))
+                    {
+                        System.IO.File.Delete(oldImagePath);
+                    }
                 }
             }
