+++ KidProjectServer/KidProjectServer/Controllers/PartyController.cs
@@ -108,10 +108,13 @@ namespace KidProjectServer.Controllers
                     await formData.Image.CopyToAsync(stream);
                 }
                 // Delete old image if it exists
-                var oldImagePath = Path.Combine(_configuration["ImagePath"], oldParty.Image);
-                if (System.IO.File.Exists(oldImagePath))
+                if(oldParty.Image != null)
                 {
-                    System.IO.File.Delete(oldImagePath);
+                    var oldImagePath = Path.Combine(_configuration["ImagePath"], oldParty.Image);
+                    if (System.IO.File.Exists(oldImagePath))
+                    {
+                        System.IO.File.Delete(oldImagePath);
+                    }
                 }
             }
